structured login and registration
- using a seperated api for api endpoints
- using controllers
